+++ src/universal/modules/userDashboard/components/EmptyOrgsCallOut/EmptyOrgsCallOut.js
@@ -21,7 +21,7 @@ const EmptyOrgsCallOut = (props) => {
   );
 
   return (
-    <CallOutPanel control={button} heading={'You don’t own any organizations!'} panelLabel={'Organizations'}>
+    <CallOutPanel control={button} heading={'You aren’t in any organizations!'} panelLabel={'Organizations'}>
       <span>{'You can create a new organization'}<br />{'and manage your own teams and tasks.'}</span>
     </CallOutPanel>
   );