+++ src/universal/types/schema.flow.js
@@ -1030,7 +1030,7 @@ export type Mutation = {
   /** Receive a webhook from github saying an org member was removed */
   githubRemoveMember: ?boolean;
   /** pauses the subscription for a single user */
-  inactivateUser: ?boolean;
+  inactivateUser: ?InactivateUserPayload;
   /** If in the org,
      Send invitation emails to a list of email addresses, add them to the invitation table.
      Else, send a request to the org leader to get them approval and put them in the OrgApproval table. */
@@ -1392,7 +1392,7 @@ export type RetrospectiveMeeting = {
   viewerMeetingMember: ?RetrospectiveMeetingMember;
   /** the threshold used to achieve the autogroup. Useful for model tuning. Serves as a flag if autogroup was used. */
   autoGroupThreshold: ?number;
-  /** the next smallest threshold to guarantee at least 1 more grouping will be achieved */
+  /** the next smallest distance threshold to guarantee at least 1 more grouping will be achieved */
   nextAutoGroupThreshold: ?number;
   /** The grouped reflections */
   reflectionGroups: Array<RetroReflectionGroup>;
@@ -1553,6 +1553,8 @@ export type DraggerCoords = {
 }
 
 export type GoogleAnalyzedEntity = {
+  /** The lemma (dictionary entry) of the entity name. Fancy way of saying the singular form of the name, if plural. */
+  lemma: string;
   /** The name of the entity. Usually 1 or 2 words. Always a noun, sometimes a proper noun. */
   name: string;
   /** The salience of the entity in the provided text. The salience of all entities always sums to 1 */
@@ -1783,6 +1785,12 @@ export type EndMeetingPayload = {
   /** The list of tasks that were archived during the meeting */
   archivedTasks: ?Array<Task>;
   meeting: ?Meeting;
+}
+
+export type InactivateUserPayload = {
+  error: ?StandardMutationError;
+  /** The user that has been inactivated */
+  user: ?User;
 }
 
 /**